add entity

+++ StockDealDal/Dto/CreateStockDetailDto.cs
@@ -13,8 +13,8 @@ namespace StockDealDal.Dto
 
         public string Description { get; set; }
 
-        [Range(0, ulong.MaxValue, ErrorMessage = "ERR_INVALID_VALUE")]
-        public ulong Quantity { get; set; }
+        [Range(0, int.MaxValue, ErrorMessage = "ERR_INVALID_VALUE")]
+        public int Quantity { get; set; }
 
         [Range(0, double.MaxValue, ErrorMessage = "ERR_INVALID_VALUE")]
         public decimal TotalPrice { get; set; }
